move select box component to ui package
