+++ packages/ui/src/components/SelectBox/SelectItem.tsx
@@ -1,0 +1,25 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import * as Select from '@radix-ui/react-select';
+import { CheckIcon } from '@radix-ui/react-icons';
+import { forwardRef } from 'react';
+import type { AnyData } from '@polkadot-live/types/misc';
+
+export const SelectItem = forwardRef(function SelectItem(
+  { children, className, ...props }: AnyData,
+  forwardedRef
+) {
+  return (
+    <Select.Item
+      className={`SelectItem ${className}`}
+      {...props}
+      ref={forwardedRef}
+    >
+      <Select.ItemText>{children}</Select.ItemText>
+      <Select.ItemIndicator className="SelectItemIndicator">
+        <CheckIcon />
+      </Select.ItemIndicator>
+    </Select.Item>
+  );
+});
