+++ packages/ui/src/components/SelectBox/index.tsx
@@ -1,0 +1,5 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export * from './SelectItem';
+export * from './Wrappers';
